Implementado
close #98
