+++ pwdj/core/templates/base.html
@@ -21,7 +21,7 @@
             <span class="navbar-toggler-icon"></span>
         </button>
         <div class="collapse navbar-collapse" id="navbarNav">
-            <ul class="navbar-nav">
+            <ul class="navbar-nav mr-auto">
                 <li class="nav-item">
                     <a class="nav-link" href="/contato/">Contato</a>
                 </li>
@@ -29,6 +29,11 @@
                     <a class="nav-link" href="{% url 'galeria:index' %}">Galeria</a>
                 </li>
             </ul>
+            {% if user.is_authenticated %}
+                <a class="btn btn-danger my-2 my-sm-2" href="{% url 'logout' %}">Sair</a>
+            {% else%}
+                <a class="btn btn-dark my-2 my-sm-2" href="{% url 'login' %}">Entrar</a>
+            {% endif %}
         </div>
   </nav>
