+++ pwdj/core/templates/registration/login.html
@@ -1,0 +1,60 @@
+{% extends 'base.html' %}
+{% block title %}Login{% endblock title %}
+{% block content %}
+    <div class="container">
+        <div class="row">
+            <div class="col">
+                {% if form.errors %}
+                    <p>Usuário e senha inválidos. Tente de novo.</p>
+                {% endif %}
+
+                {% if next %}
+                    {% if user.is_authenticated %}
+                        <p>Sua conta não pode acessar essa página. Para prosseguir, acesse com uma
+                            conta que tenha permissão.</p>
+                    {% else %}
+                        <p>Faça login para acessar essa página.</p>
+                    {% endif %}
+                {% endif %}
+
+                <form method="post" action="{% url 'login' %}">
+                    {% csrf_token %}
+                    <form role="form">
+                        <div class="form-group">
+                            <label for="{{ form.username.id_for_label }}">
+                                {{ form.username.label }}
+                            </label>
+                            <input type="email" name="{{ form.username.name }}"
+                                   class="form-control"
+                                   id="{{ form.username.id_for_label }}"
+                                   {% if form.username.value %}
+                                    value="{{form.username.value }}"
+                                   {% endif %}
+
+                                   required="required"/>
+                        </div>
+                        <div class="form-group">
+                            <label for="exampleInputEmail1" for="{{ form.password.id_for_label }}">
+                                {{ form.password.label }}
+                            </label>
+                            <input type="password" name="{{ form.password.name }}"
+                                   class="form-control"
+                                   id="{{ form.password.id_for_label }}"
+                                   required="required"/>
+                        </div>
+                        <button type="submit" class="btn btn-primary">
+                            Login
+                        </button>
+                        <input type="hidden" name="next" value="{{ next }}"/>
+                    </form>
+                </form>
+                {# Assumes you setup the password_reset view in your URLconf #}
+                <p><a href="{% url 'password_reset' %}">Esqueceu sua senha?</a></p>
+
+            </div>
+        </div>
+
+    </div>
+
+{% endblock content %}
+
