+++ pwdj/galeria/tests/test_index_view.py
